Improve cQASM reader test
