+++ tests/golden/test_cqasm_conditions.qasm
@@ -1,0 +1,38 @@
+version 1.0
+# this file has been automatically generated by the OpenQL compiler please do not modify it manually.
+qubits 7
+
+._0
+    x q[1]
+    cond(b[8]) z q[1]
+    y q[1]
+    cond(!b[8]) z q[1]
+    x q[1]
+    cond(b[8]) z q[1]
+    cond(b[8]&&b[7) x q[1]
+    cond(b[8]) y q[1]
+    cond(b[7]) x q[1]
+    cond(b[8]||b[7) z q[1]
+    x q[1]
+    cond(b[8]) y q[1]
+    z q[1]
+    cond(b[7]) x q[1]
+    cond(b[8]^^b[7) y q[1]
+    cond(!b[8]) z q[1]
+    cond(b[8]) x q[1]
+    cond(!b[7]) y q[1]
+    cond(b[7]) z q[1]
+    cond(!(b[8]&&b[7)) x q[1]
+    cond(!b[8]) y q[1]
+    z q[1]
+    cond(!b[7]) x q[1]
+    y q[1]
+    cond(!(b[8]||b[7)) z q[1]
+    cond(!b[8]) y q[1]
+    cond(!b[7]) x q[1]
+    cond(!(b[8]^^b[7)) y q[1]
+    cond(b[8]) z q[1]
+    cond(!b[8]) x q[1]
+    cond(b[7]) y q[1]
+    cond(!b[7]) z q[1]
+    wait 1
