+++ tests/golden/test_do_while.qisa
@@ -10,16 +10,16 @@ smis s8, {0, 1, 5, 6}
 smis s9, {2, 3, 4} 
 start:
 
-aKernel1_do_while0_start:
+aKernel1_do_while2_start:
 
 aKernel1:
     1    x s0
     qwait 2
 
-aKernel1_do_while0:
+aKernel1_do_while2:
     cmp r2, r3
     nop
-    br lt, aKernel1_do_while0_start
+    br lt, aKernel1_do_while2_start
 
 aKernel2:
     1    y s0
