+++ tests/golden/test_multiple_programs.qasm
@@ -1,0 +1,18 @@
+version 1.0
+# this file has been automatically generated by the OpenQL compiler please do not modify it manually.
+qubits 7
+
+._0
+    prepz q[0]
+    prepz q[1]
+    x q[0]
+    h q[1]
+    wait 1
+
+._1
+    prepz q[2]
+    prepz q[3]
+    x q[2]
+    h q[3]
+    { measz q[0] | measz q[1] | measz q[2] | measz q[3] | measz q[4] | measz q[5] }
+    wait 1
