+++ tests/golden/test_single_bit_kernel_operations.qasm
@@ -1,0 +1,22 @@
+version 1.0
+# this file has been automatically generated by the OpenQL compiler please do not modify it manually.
+qubits 7
+
+._0
+    { prepz q[0] | prepz q[1] | prepz q[2] | prepz q[3] }
+    prepz q[4]
+    { i q[0] | i q[1] | i q[2] | i q[3] }
+    h q[4]
+    { x q[1] | x q[2] | x q[3] }
+    { z q[0] | z q[1] }
+    s q[1]
+    { sdag q[2] | sdag q[3] }
+    { tdag q[1] | tdag q[2] | tdag q[3] }
+    rx90 q[3]
+    xm90 q[1]
+    ry90 q[0]
+    { ym90 q[2] | ym90 q[3] }
+    { measz q[2] | measz q[3] }
+    { measz q[0] | measz q[1] }
+    measz q[4]
+    wait 1
