+++ tests/golden/test_sub_circuit_programs.qasm
@@ -1,0 +1,26 @@
+version 1.0
+# this file has been automatically generated by the OpenQL compiler please do not modify it manually.
+qubits 7
+
+.init_0
+    prepz q[0]
+    prepz q[1]
+    wait 1
+
+.do_somework_1_for1_start
+    ldi r29, 3
+    ldi r30, 1
+    ldi r31, 0
+
+.do_somework_1
+    x q[0]
+    h q[1]
+    wait 1
+
+.do_somework_1_for1_end
+    add r31, r31, r30
+    blt r31, r29, do
+
+.do_measurement_2
+    { measz q[0] | measz q[1] | measz q[2] | measz q[3] | measz q[4] | measz q[5] }
+    wait 1
